backend/requirements.txt
python-docx
docx2txt
pdfplumber
google-generativeai
click